Remove business logic as much as quickly possible from Embed.tsx
