+++ packages/features/embed/lib/getDimension.tsx
@@ -1,0 +1,6 @@
+export const getDimension = (dimension: string) => {
+  if (dimension.match(/^\d+$/)) {
+    dimension = `${dimension}%`;
+  }
+  return dimension;
+};
